requirements.txt
streamlit==1.21.0
langchain==0.15.0
langchain_groq==0.1.0
langchain_community==0.1.0
python-dotenv==1.0.0
requests==2.28.2
duckduckgo-search==0.5.0